Modernize CollabVMServer by not using io_service
io_service is a legacy typedef and ideally the plan is to disable all
deprecated ASIO features. (this will be further painful when we get into
removing io_service::work usage, however the bulk of that has been done
and there's basically only one usage of it now.)

Also makes the main() a bit nicer by refactoring the argument code.