CMake: Do not allow in-tree builds
Adds COLLABVM_BUILD_FEATURES.
This is a CMake list type, which can be specified on the
command line to add flavors. Any target flavored with
collabvm_targetize automatically gets support on the fly.

Adds support for ASAN and TSAN instrumentation,
using COLLABVM_BUILD_FEATURES.

Moves the plugin ABI into include/plugins/ as that will make it
a bit easier to install.